Update compojure version to latest
Changing the version number to 1.1.6, the latest.  This solves an issue where 1.1.5 pulls in an OLD version of ring/ring-core.  This causes issues with other common dependencies like cemerick.friend.  The alternative is to explicitly specify `[ring/ring-core "1.2.0"]` as a dependency.

I made the change locally, installed the template and was able to `lein new compojure-app` and get a working repl and server.